SIH/sih/sih/requirements.txt
streamlit>=1.28.0
Pillow>=9.0.0
requests>=2.28.0
